+++ CHANGELOG.md
@@ -7,6 +7,46 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 # Unreleased
 
+- **fixed:** Fully qualify `FromReflect` in generated code ([#107])
+
+[#107]: https://github.com/EmbarkStudios/mirror-mirror/pull/107
+
+# 0.1.9 (24. February, 2023)
+
+- **added:** Add `StructValue::with_capacity`,
+  `TupleStructValue::with_capacity`, and `TupleValue::with_capacity` ([#106])
+- **added:** Add `EnumValue::new_struct_variant_with_capacity` and
+  `EnumValue::new_struct_variant_with_capacity` constructors ([#106])
+- **fixed:** In `Reflect::to_value` for enums, only generate a catch all branch
+  if the enum has a variant with `#[reflect(skip)]` ([#105])
+- **added:** Add a `has_default_value` method to types in `type_info` ([#104])
+
+[#105]: https://github.com/EmbarkStudios/mirror-mirror/pull/105
+[#104]: https://github.com/EmbarkStudios/mirror-mirror/pull/104
+[#106]: https://github.com/EmbarkStudios/mirror-mirror/pull/106
+
+# 0.1.8 (23. February, 2023)
+
+- **added:** Make `Key` and `KeyPath` impl `Hash` ([#103])
+
+[#103]: https://github.com/EmbarkStudios/mirror-mirror/pull/103
+
+# 0.1.7 (23. February, 2023)
+
+- **added:** Implement `PartialEq`, `Eq`, `Hash` for types in `type_info` ([#100] [#101])
+
+[#100]: https://github.com/EmbarkStudios/mirror-mirror/pull/100
+[#101]: https://github.com/EmbarkStudios/mirror-mirror/pull/101
+
+# 0.1.6 (16. February, 2023)
+
+- **fixed:** Fix inconsistent ordering when iterating over fields in struct
+  values and struct types. Same for struct variants ([#98])
+
+[#98]: https://github.com/EmbarkStudios/mirror-mirror/pull/98
+
+# 0.1.5 (14. February, 2023)
+
 - **added:** Add visitor API ([#92])
 - **added:** Add `fields_len` methods to the following types ([#94])
     - `StructType`
@@ -16,15 +56,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     - `StructVariant`
     - `TupleStructVariant`
 - **added:** Add `EnumType::variants_len` ([#94])
+- **added:** Support setting a default value for `OpaqueType` ([#97])
+- **added:** Support pretty printing types ([#95])
 
 [#92]: https://github.com/EmbarkStudios/mirror-mirror/pull/92
-[#93]: https://github.com/EmbarkStudios/mirror-mirror/pull/93
 [#94]: https://github.com/EmbarkStudios/mirror-mirror/pull/94
+[#95]: https://github.com/EmbarkStudios/mirror-mirror/pull/95
+[#97]: https://github.com/EmbarkStudios/mirror-mirror/pull/97
 
-# 0.1.4 (08. February, 2023)
+# 0.1.4 (13. February, 2023)
 
 - **added:** Implement `Hash` to `Value` ([#93])
 
+[#93]: https://github.com/EmbarkStudios/mirror-mirror/pull/93
 
 # 0.1.3 (08. February, 2023)